ep1: Reverted Float back to Int. Next TDD objective: Add curl in loop to validate reward level range A-J when purchaseTotal from 100 to 1000

+++ src/main/kotlin/com/urbn/rewards/models/Customer.kt
@@ -2,7 +2,7 @@ package com.urbn.rewards.models
 
 data class Customer(
     val email: String,
-    val rewardPoints: Float,
+    val rewardPoints: Int,
     val rewardsTier: String,
     val nextRewardsTier: String,
     val nextRewardsTierName: String,
