+++ src/main/kotlin/com/urbn/rewards/models/OrderRequest.kt
@@ -2,5 +2,5 @@ package com.urbn.rewards.models
 
 data class OrderRequest(
     val email: String,
-    val purchaseTotal: Float
+    val purchaseTotal: Int
 )
